move pre-project loading events (A) to core
