+++ core/dbt/events/types.py
@@ -1,5 +1,204 @@
-from dbt.events.base_types import WarnLevel
+from dbt.events.base_types import WarnLevel, InfoLevel, DebugLevel, ErrorLevel
+
 from dbt.common.ui import warning_tag, line_wrap_message
+
+
+# =======================================================
+# A - Pre-project loading
+# =======================================================
+
+
+class MainReportVersion(InfoLevel):
+    def code(self) -> str:
+        return "A001"
+
+    def message(self) -> str:
+        return f"Running with dbt{self.version}"
+
+
+class MainReportArgs(DebugLevel):
+    def code(self) -> str:
+        return "A002"
+
+    def message(self) -> str:
+        return f"running dbt with arguments {str(self.args)}"
+
+
+class MainTrackingUserState(DebugLevel):
+    def code(self) -> str:
+        return "A003"
+
+    def message(self) -> str:
+        return f"Tracking: {self.user_state}"
+
+
+class MergedFromState(DebugLevel):
+    def code(self) -> str:
+        return "A004"
+
+    def message(self) -> str:
+        return f"Merged {self.num_merged} items from state (sample: {self.sample})"
+
+
+class MissingProfileTarget(InfoLevel):
+    def code(self) -> str:
+        return "A005"
+
+    def message(self) -> str:
+        return f"target not specified in profile '{self.profile_name}', using '{self.target_name}'"
+
+
+# Skipped A006, A007
+
+
+class InvalidOptionYAML(ErrorLevel):
+    def code(self) -> str:
+        return "A008"
+
+    def message(self) -> str:
+        return f"The YAML provided in the --{self.option_name} argument is not valid."
+
+
+class LogDbtProjectError(ErrorLevel):
+    def code(self) -> str:
+        return "A009"
+
+    def message(self) -> str:
+        msg = "Encountered an error while reading the project:"
+        if self.exc:
+            msg += f"  ERROR: {str(self.exc)}"
+        return msg
+
+
+# Skipped A010
+
+
+class LogDbtProfileError(ErrorLevel):
+    def code(self) -> str:
+        return "A011"
+
+    def message(self) -> str:
+        msg = "Encountered an error while reading profiles:\n" f"  ERROR: {str(self.exc)}"
+        if self.profiles:
+            msg += "Defined profiles:\n"
+            for profile in self.profiles:
+                msg += f" - {profile}"
+        else:
+            msg += "There are no profiles defined in your profiles.yml file"
+
+        msg += """
+For more information on configuring profiles, please consult the dbt docs:
+
+https://docs.getdbt.com/docs/configure-your-profile
+"""
+        return msg
+
+
+class StarterProjectPath(DebugLevel):
+    def code(self) -> str:
+        return "A017"
+
+    def message(self) -> str:
+        return f"Starter project path: {self.dir}"
+
+
+class ConfigFolderDirectory(InfoLevel):
+    def code(self) -> str:
+        return "A018"
+
+    def message(self) -> str:
+        return f"Creating dbt configuration folder at {self.dir}"
+
+
+class NoSampleProfileFound(InfoLevel):
+    def code(self) -> str:
+        return "A019"
+
+    def message(self) -> str:
+        return f"No sample profile found for {self.adapter}."
+
+
+class ProfileWrittenWithSample(InfoLevel):
+    def code(self) -> str:
+        return "A020"
+
+    def message(self) -> str:
+        return (
+            f"Profile {self.name} written to {self.path} "
+            "using target's sample configuration. Once updated, you'll be able to "
+            "start developing with dbt."
+        )
+
+
+class ProfileWrittenWithTargetTemplateYAML(InfoLevel):
+    def code(self) -> str:
+        return "A021"
+
+    def message(self) -> str:
+        return (
+            f"Profile {self.name} written to {self.path} using target's "
+            "profile_template.yml and your supplied values. Run 'dbt debug' to "
+            "validate the connection."
+        )
+
+
+class ProfileWrittenWithProjectTemplateYAML(InfoLevel):
+    def code(self) -> str:
+        return "A022"
+
+    def message(self) -> str:
+        return (
+            f"Profile {self.name} written to {self.path} using project's "
+            "profile_template.yml and your supplied values. Run 'dbt debug' to "
+            "validate the connection."
+        )
+
+
+class SettingUpProfile(InfoLevel):
+    def code(self) -> str:
+        return "A023"
+
+    def message(self) -> str:
+        return "Setting up your profile."
+
+
+class InvalidProfileTemplateYAML(InfoLevel):
+    def code(self) -> str:
+        return "A024"
+
+    def message(self) -> str:
+        return "Invalid profile_template.yml in project."
+
+
+class ProjectNameAlreadyExists(InfoLevel):
+    def code(self) -> str:
+        return "A025"
+
+    def message(self) -> str:
+        return f"A project called {self.name} already exists here."
+
+
+class ProjectCreated(InfoLevel):
+    def code(self) -> str:
+        return "A026"
+
+    def message(self) -> str:
+        return f"""
+Your new dbt project "{self.project_name}" was created!
+
+For more information on how to configure the profiles.yml file,
+please consult the dbt documentation here:
+
+  {self.docs_url}
+
+One more thing:
+
+Need help? Don't hesitate to reach out to us via GitHub issues or on Slack:
+
+  {self.slack_url}
+
+Happy modeling!
+"""
 
 
 # =======================================================
