+++ core/dbt/task/init.py
@@ -19,7 +19,7 @@ from dbt.adapters.factory import load_plugin, get_include_paths
 from dbt.contracts.util import Identifier as ProjectName
 
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.events.types import (
     StarterProjectPath,
     ConfigFolderDirectory,
     NoSampleProfileFound,
