+++ tests/unit/test_events.py
@@ -120,24 +120,24 @@ sample_values = [
     # N.B. Events instantiated here include the module prefix in order to
     # avoid having the entire list twice in the code.
     # A - pre-project loading
-    types.MainReportVersion(version=""),
-    types.MainReportArgs(args={}),
-    types.MainTrackingUserState(user_state=""),
-    types.MergedFromState(num_merged=0, sample=[]),
-    types.MissingProfileTarget(profile_name="", target_name=""),
-    types.InvalidOptionYAML(option_name="vars"),
-    types.LogDbtProjectError(),
-    types.LogDbtProfileError(),
-    types.StarterProjectPath(dir=""),
-    types.ConfigFolderDirectory(dir=""),
-    types.NoSampleProfileFound(adapter=""),
-    types.ProfileWrittenWithSample(name="", path=""),
-    types.ProfileWrittenWithTargetTemplateYAML(name="", path=""),
-    types.ProfileWrittenWithProjectTemplateYAML(name="", path=""),
-    types.SettingUpProfile(),
-    types.InvalidProfileTemplateYAML(),
-    types.ProjectNameAlreadyExists(name=""),
-    types.ProjectCreated(project_name=""),
+    core_types.MainReportVersion(version=""),
+    core_types.MainReportArgs(args={}),
+    core_types.MainTrackingUserState(user_state=""),
+    core_types.MergedFromState(num_merged=0, sample=[]),
+    core_types.MissingProfileTarget(profile_name="", target_name=""),
+    core_types.InvalidOptionYAML(option_name="vars"),
+    core_types.LogDbtProjectError(),
+    core_types.LogDbtProfileError(),
+    core_types.StarterProjectPath(dir=""),
+    core_types.ConfigFolderDirectory(dir=""),
+    core_types.NoSampleProfileFound(adapter=""),
+    core_types.ProfileWrittenWithSample(name="", path=""),
+    core_types.ProfileWrittenWithTargetTemplateYAML(name="", path=""),
+    core_types.ProfileWrittenWithProjectTemplateYAML(name="", path=""),
+    core_types.SettingUpProfile(),
+    core_types.InvalidProfileTemplateYAML(),
+    core_types.ProjectNameAlreadyExists(name=""),
+    core_types.ProjectCreated(project_name=""),
     # D - Deprecations ======================
     core_types.PackageRedirectDeprecation(old_name="", new_name=""),
     core_types.PackageInstallPathDeprecation(),
